backend/requirements.txt
Flask
python-socketio
Flask-SocketIO
eventlet
# or instead of Flask-SocketIO and eventlet, use:
# websockets
# if using FastAPI:
# fastapi
# uvicorn[standard]

opencv-python
Pillow
qrcode[pil]
numpy
openai-whisper
llama-cpp-python

# Add other AI/ML libraries later, e.g.:
# transformers
# torch
# accelerate
# bitsandbytes